Transcript View: Add Transcript View Config, Music Sections, TypeScript Declarations, and Fix Scrolling (#321)
This re-adds the TranscriptView config to wrap the transcript entries array. This is useful for passing additional data into transcript without adding a bunch more properties.
This also adds support for music sections where we skip parts of the program because music is playing.
It also fixes some scroll issues, as well as adds a TypeScript declarations file so other TS projects can import it easily.