Transcript Scrolling Fix (#395)
Fix a scrolling bug and reduce CPU use

We were incorrectly comparing whether the entries had changed so it was always updating the current entries. Now it only updates when they change. This was causing scroll stuttering when you clicked the "Scroll Text with Audio" because it was being updated too frequently.