Allow the SearchHandler to be injected into the Radio Player
In order to use a search engine external to the radio player, we need to be able to inject a search handler into the radio player configured with appropriately for the environment.

This adds a `SearchHandlerInterface` and exposes a `searchHandler` property on the radio player to inject it. We can then use any search engine that conforms to the interface. We currently have the one `LocalSearchIndex` that does client-side search and we will add a `RadioArchiveSearchIndex` that uses the backend search service.