Enable Hosted Fields from the Start
Previously, we were rendering the hosted fields when the user clicked on "Credit Card", but they are kind of slow to render, over a second on a fast machine/connection.

Now we render them immediately, but leave them hidden. The initial delay isn't noticeable because they are hidden during the initial load.