Refactor logo/homepage link for a11y
To comply with WCAG 2.2 Success Criterion 1.1.1 Non-text Content,
specifically decoration the svg logo itself needs an aria-hidden="true"
attribute (focusable=false is added for IE compatibility) so that it
can be ignored by assistive technology as it adds no useful context.
The only thing necessary is an aria-label on the anchor tag.